moved chip & chip-input to their own Chips  group
