+++ projects/ui-framework/src/lib/buttons-indicators/avatar/avatar.component.spec.ts
@@ -5,8 +5,8 @@ import { By } from '@angular/platform-browser';
 import { IconsModule } from '../../icons/icons.module';
 import { DOMhelpers } from '../../services/utils/dom-helpers.service';
 import { SimpleChange } from '@angular/core';
-import { ChipType } from '../chip/chip.enum';
-import { ChipModule } from '../chip/chip.module';
+import { ChipType } from '../../chips/chip/chip.enum';
+import { ChipModule } from '../../chips/chip/chip.module';
 import { TruncateTooltipModule } from '../../services/truncate-tooltip/truncate-tooltip.module';
 import { BadgeSize, AvatarBadges } from './avatar.consts';
