+++ projects/ui-framework/src/lib/buttons-indicators/avatar/avatar.interface.ts
@@ -1,4 +1,4 @@
-import { Chip } from '../chip/chip.interface';
+import { Chip } from '../../chips/chip/chip.interface';
 import { AvatarBadge, AvatarOrientation } from './avatar.enum';
 
 export interface BadgeConfig {
