+++ projects/ui-framework/src/lib/buttons-indicators/avatar/avatar.stories.ts
@@ -13,7 +13,7 @@ import { StoryBookLayoutModule } from '../../story-book-layout/story-book-layout
 import { AvatarModule } from './avatar.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { IconsModule } from '../../icons/icons.module';
-import { ChipType } from '../chip/chip.enum';
+import { ChipType } from '../../chips/chip/chip.enum';
 
 const avatarStories = storiesOf(
   `${ComponentGroupType.ButtonsAndIndicators}.Avatar`,
