+++ projects/ui-framework/src/lib/chips/chip-input/chip-input.component.scss
