+++ projects/ui-framework/src/lib/chips/chip-input/chip-input.interface.ts
@@ -1,4 +1,4 @@
-import { InputEventType } from '../form-elements.enum';
+import { InputEventType } from '../../form-elements/form-elements.enum';
 
 export interface ChipInputChange {
   value: string[];
