+++ projects/ui-framework/src/lib/chips/chip-input/chip-input.module.ts
@@ -4,7 +4,7 @@ import { ChipInputComponent } from './chip-input.component';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { IconsModule } from '../../icons/icons.module';
 import { InputMessageModule } from '../../form-elements/input-message/input-message.module';
-import { ChipModule } from '../../buttons-indicators/chip/chip.module';
+import { ChipModule } from '../chip/chip.module';
 
 @NgModule({
   declarations: [ChipInputComponent],
