+++ projects/ui-framework/src/lib/chips/chip/chip.component.scss
