+++ projects/ui-framework/src/lib/chips/chip/chip.enum.ts
