+++ projects/ui-framework/src/lib/chips/chip/chip.module.ts
