+++ projects/ui-framework/src/lib/chips/chip/chips.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { ChipComponent } from './chip.component';
-import { ChipInputComponent } from '../../form-elements/chip-input/chip-input.component';
+import { ChipInputComponent } from '../chip-input/chip-input.component';
 import { ChipModule } from './chip.module';
-import { ChipInputModule } from '../../form-elements/chip-input/chip-input.module';
+import { ChipInputModule } from '../chip-input/chip-input.module';
 
 @NgModule({
   declarations: [],
